Giving some structure to Test Cluster state for walrus-sites
