+++ site-builder/Cargo.toml
@@ -42,3 +42,5 @@ tracing-subscriber = "0.3.18"
 sui-move-build = { git = "https://github.com/MystenLabs/sui", tag = "testnet-v1.49.1" }
 walrus-service = { git = "https://github.com/MystenLabs/walrus", tag = "testnet-v1.26.2", features = ["test-utils"] }
 walrus-sui = { git = "https://github.com/MystenLabs/walrus", tag = "testnet-v1.26.2" }
+walrus-sdk = { git = "https://github.com/MystenLabs/walrus", tag = "testnet-v1.26.2" }
+walrus-test-utils = { git = "https://github.com/MystenLabs/walrus", tag = "testnet-v1.26.2" }
